Tests: move SIMD complex tests to dedicated file and expand coverage
