+++ Tests/CMakeLists.txt
@@ -19,6 +19,7 @@ include(GoogleTest)
 
 set(SOURCES
         base/SIMD_test.cpp
+        base/SIMD/SIMD_complex_test.cpp
         core/Phase_test.cpp
         core/Expected_test.cpp
         core/AudioBuffer_test.cpp
